% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLN_and_CLT.R
\name{CLT}
\alias{CLT}
\title{Verify the Central Limit Theorem by figure}
\usage{
CLT(r, parameter, mean, var, n, N = 1000)
}
\arguments{
\item{r}{the distribution}

\item{parameter}{the parameter of this distribution}

\item{mean}{the mean of this distribution}

\item{var}{the variance of this distribution}

\item{n}{the number of samples}

\item{N}{the repetition times}
}
\value{
the figure
}
\description{
Verify the Central Limit Theorem by figure
}
\examples{
\dontrun{
n <- c(1,10,30,100)
par(mfrow=c(4,4))
CLT(rbinom, parameter = c(1, 0.3), 0.3, 0.21, n)
CLT(rpois, parameter = 1, 1, 1, n)
CLT(runif, parameter = c(0,1), 0.5, 1/12, n)
CLT(rexp, parameter = 1, 1, 1, n)
}
}
