% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complex_det.R
\name{Complex_det}
\alias{Complex_det}
\title{The determinant of a complex matrix}
\usage{
Complex_det(x)
}
\arguments{
\item{x}{the complex matrix}
}
\value{
the determinant of x\code{n}
}
\description{
The determinant of a complex matrix
}
\examples{
\dontrun{
x <- matrix(c(1+1i,2,3+4i,2i,3+7i,1-3i,2+4i,-1i,9),3,3)
Complex_det(x)
}
}
