% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_choose.R
\name{k_choose1}
\alias{k_choose1}
\title{Selecting the optimal ridge regression parameters by GCV}
\usage{
k_choose1(X, Y, kmax, N)
}
\arguments{
\item{X}{the independent variable}

\item{Y}{the response variable}

\item{kmax}{the maximum value of the ridge regression parameter}

\item{N}{the number of cross validation}
}
\value{
the optimal ridge regression parameter
}
\description{
Selecting the optimal ridge regression parameters by GCV
}
\examples{
\dontrun{
library(MASS)
data <- read.table("E:/Rcpp/data.txt", header = TRUE)
x <- as.matrix(data[,2:21])
y <- data[,1]
k_best1 <- k_choose1(x, y, kmax = 2, N = 2000)
print(k_best1)
 #compare with the optimal parameter gotten by existing ridge package 
select(lm.ridge(y~x, data = data, lambda = seq(0,2,0.001)))}
}
