% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_choose.R
\name{k_choose2}
\alias{k_choose2}
\title{Selecting the optimal ridge regression parameters by minimuming GIC}
\usage{
k_choose2(X, Y)
}
\arguments{
\item{X}{the independent variable}

\item{Y}{the response variable}
}
\value{
the optimal ridge regression parameter
}
\description{
Getting the optimal ridge regression parameters by minimuming GIC
}
\examples{
\dontrun{
data <- read.table("E:/Rcpp/data.txt", header = TRUE)
x <- as.matrix(data[,2:21])
y <- data[,1]
k_best2 <- k_choose2(x, y)
}
}
