% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLN_and_CLT.R
\name{LLN}
\alias{LLN}
\title{Verify the Law of Large Numbers by figure}
\usage{
LLN(r, mean, n)
}
\arguments{
\item{r}{the distribution}

\item{mean}{the mean of this distribution}

\item{n}{the number of samples}
}
\value{
the figure
}
\description{
Verify the law of large Numbers by figure
}
\examples{
\dontrun{
n <- 1000
LLN(rbinom(n, 1, 0.3), 0.3, n)
LLN(runif(n,0,1),0.5,n)
LLN(rpois(n, 1), 1, n)
LLN(rexp(n, 1), 1, n)
}
}
