% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwC}
\alias{rwC}
\title{Cpp function for Exercise 9.4}
\usage{
rwC(sigma, x_0, N)
}
\arguments{
\item{sigma}{the variance of the normality distribution}

\item{x_0}{the original value}

\item{N}{the size}
}
\value{
the sample x \code{n}
}
\description{
Cpp function for Exercise 9.4
}
