% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_choose.R
\name{k_choose2}
\alias{k_choose2}
\title{Selecting the optimal ridge regression parameters by minimuming GIC}
\usage{
k_choose2(X, Y)
}
\arguments{
\item{X}{the independent variable}

\item{Y}{the response variable}
}
\value{
the optimal ridge regression parameter
}
\description{
Getting the optimal ridge regression parameters by minimuming GIC
}
